Fix migration impact for custom Shell renderer users
