+++ src/Controls/src/Core/Handlers/Shell/ShellSectionHandler.Android.cs
@@ -3,8 +3,10 @@ using System;
 using System.Collections.Generic;
 using System.Collections.Specialized;
 using System.Linq;
+using Android.Content;
 using Android.OS;
 using Android.Views;
+using Android.Views.Animations;
 using Android.Widget;
 using AndroidX.Fragment.App;
 using AndroidX.ViewPager2.Adapter;
@@ -13,6 +15,7 @@ using Google.Android.Material.Tabs;
 using Microsoft.Maui.Controls.Internals;
 using Microsoft.Maui.Controls.Platform.Compatibility;
 using Microsoft.Maui.Graphics;
+using AAnimation = Android.Views.Animations.Animation;
 using AView = Android.Views.View;
 using LP = Android.Views.ViewGroup.LayoutParams;
 
@@ -159,6 +162,24 @@ namespace Microsoft.Maui.Controls.Handlers
 
             return rootView;
         }
+
+        /// <summary>
+        /// Creates the animation for a navigation transition within this Shell section.
+        /// Override to provide custom page transition animations for push/pop navigation.
+        /// Return null to use the default navigation animation.
+        /// </summary>
+        /// <param name="context">The Android context for loading animation resources.</param>
+        /// <param name="isPopping">True if navigating back, false if navigating forward.</param>
+        /// <param name="enter">True for the entering page's animation, false for the exiting page's animation.</param>
+        /// <returns>The animation to use, null to use default, or an animation with zero duration to skip animation.</returns>
+        protected virtual AAnimation? OnCreateNavigationAnimation(Context context, bool isPopping, bool enter)
+        {
+            return null;
+        }
+
+        // Internal accessor for ShellStackNavigationManager to invoke the protected virtual
+        internal AAnimation? InvokeOnCreateNavigationAnimation(Context context, bool isPopping, bool enter)
+            => OnCreateNavigationAnimation(context, isPopping, enter);
 
         protected override void ConnectHandler(AView platformView)
         {
@@ -851,543 +872,6 @@ namespace Microsoft.Maui.Controls.Handlers
     }
 
     /// <summary>
-    /// Fragment that hosts a ShellContent page with its own StackNavigationManager.
-    /// This enables each content tab to have independent navigation.
-    /// </summary>
-    internal class ShellContentNavigationFragment : Fragment, IStackNavigation
-    {
-        ShellContent? _shellContent;
-        IMauiContext? _mauiContext;
-        ShellSectionHandler? _handler;
-        StackNavigationManager? _stackNavigationManager;
-        FragmentContainerView? _navigationContainer;
-        Page? _rootPage;
-        int _navigationContainerId;
-        ShellContentStackNavigationView? _navigationViewAdapter;
-
-        // Default constructor required by Android's FragmentManager for fragment restoration.
-        // When FragmentStateAdapter (ViewPager2) saves and restores fragment state during tab switches,
-        // it uses Fragment.instantiate() which requires a parameterless constructor.
-        public ShellContentNavigationFragment()
-        {
-        }
-
-        public ShellContentNavigationFragment(ShellContent? shellContent, IMauiContext? mauiContext, ShellSectionHandler? handler)
-        {
-            _shellContent = shellContent;
-            _mauiContext = mauiContext;
-            _handler = handler;
-        }
-
-        public override void OnCreate(Bundle? savedInstanceState)
-        {
-            // Always pass null to prevent restoring stale child fragment state.
-            // OffscreenPageLimit keeps fragments alive so restoration shouldn't occur,
-            // but this is defense-in-depth.
-            base.OnCreate(null);
-        }
-
-        public override AView OnCreateView(LayoutInflater inflater, ViewGroup? container, Bundle? savedInstanceState)
-        {
-            // If this fragment was restored by FragmentManager without proper data,
-            // return an empty view. The ViewPager2 adapter will recreate it properly.
-            if (_shellContent is null || _mauiContext is null)
-            {
-                return new FrameLayout(inflater.Context!);
-            }
-
-            if (_navigationContainer is not null)
-            {
-                // Check if NavHostFragment needs recreation
-                var existingNavHost = ChildFragmentManager.FindFragmentById(_navigationContainerId);
-                if (existingNavHost is null && _stackNavigationManager is not null)
-                {
-                    var recreatedNavHost = new MauiNavHostFragment()
-                    {
-                        StackNavigationManager = _stackNavigationManager
-                    };
-                    ChildFragmentManager
-                        .BeginTransactionEx()
-                        .AddEx(_navigationContainerId, recreatedNavHost)
-                        .CommitNowAllowingStateLoss();
-                }
-
-                return _navigationContainer;
-            }
-
-            // Get the root page for this ShellContent
-            _rootPage = ((IShellContentController)_shellContent)?.GetOrCreateContent();
-            if (_rootPage is null)
-            {
-                return new FrameLayout(inflater.Context!);
-            }
-
-            // Subscribe to navigation events EARLY - before anything else
-            // This ensures we catch any navigation requests that come before the view is attached
-            var shellSection = _shellContent?.Parent as ShellSection;
-            if (shellSection is not null && !_subscribedToNavigationRequested)
-            {
-                ((IShellSectionController)shellSection).NavigationRequested += OnNavigationRequested;
-                _subscribedToNavigationRequested = true;
-            }
-
-            // Create FragmentContainerView for navigation stack
-            if (_navigationContainerId == 0)
-            {
-                _navigationContainerId = AView.GenerateViewId();
-            }
-
-            _navigationContainer = new FragmentContainerView(_mauiContext!.Context!)
-            {
-                Id = _navigationContainerId,
-                LayoutParameters = new LP(LP.MatchParent, LP.MatchParent)
-            };
-
-            // Create StackNavigationManager with scoped context
-            var scopedContext = _mauiContext.MakeScoped(fragmentManager: ChildFragmentManager);
-            _stackNavigationManager = new StackNavigationManager(scopedContext);
-
-            // Create NavHostFragment
-            var navHostFragment = new MauiNavHostFragment()
-            {
-                StackNavigationManager = _stackNavigationManager
-            };
-
-            ChildFragmentManager
-                .BeginTransactionEx()
-                .AddEx(_navigationContainerId, navHostFragment)
-                .CommitNowAllowingStateLoss();
-
-            _navigationContainer.ViewAttachedToWindow += OnNavigationContainerAttached;
-
-            if (_navigationContainer.IsAttachedToWindow)
-            {
-                ConnectAndInitialize();
-            }
-
-            return _navigationContainer;
-        }
-
-        bool _subscribedToNavigationRequested;
-
-        void OnNavigationContainerAttached(object? sender, AView.ViewAttachedToWindowEventArgs e)
-        {
-            _navigationContainer?.ViewAttachedToWindow -= OnNavigationContainerAttached;
-            ConnectAndInitialize();
-        }
-
-        void ConnectAndInitialize()
-        {
-            if (_stackNavigationManager is null || _rootPage is null)
-            {
-                return;
-            }
-
-            // Create adapter that wraps the root page and delegates NavigationFinished to this fragment
-            _navigationViewAdapter = new ShellContentStackNavigationView(_rootPage, this);
-
-            // Connect using the adapter (which properly implements IStackNavigationView via page delegation)
-            _stackNavigationManager.Connect(_navigationViewAdapter, _navigationContainer);
-
-            // Apply dark/light background to the navigation container when the page has no explicit
-            // Background, matching old ShellPageContainer constructor behavior.
-            // We set it on the container (not the page's platform view) because the page's handler
-            // hasn't been created yet at this point — StackNavigationManager creates it asynchronously.
-            // The page view is transparent by default, so the container background shows through.
-            if (_rootPage is IView view && view.Background is null && _navigationContainer is not null)
-            {
-                var context = _mauiContext!.Context!;
-                bool isDark = Controls.Application.Current?.RequestedTheme == ApplicationModel.AppTheme.Dark;
-                int bgColor = isDark
-                    ? AndroidX.Core.Content.ContextCompat.GetColor(context, global::Android.Resource.Color.BackgroundDark)
-                    : AndroidX.Core.Content.ContextCompat.GetColor(context, global::Android.Resource.Color.BackgroundLight);
-                _navigationContainer.SetBackgroundColor(new global::Android.Graphics.Color(bgColor));
-            }
-
-            // Subscribe to navigation events if not already subscribed
-            // (may have already been done in OnCreateView)
-            var shellSection = _shellContent?.Parent as ShellSection;
-            if (shellSection is not null && !_subscribedToNavigationRequested)
-            {
-                ((IShellSectionController)shellSection).NavigationRequested += OnNavigationRequested;
-                _subscribedToNavigationRequested = true;
-            }
-
-            // Build the initial stack from the ShellSection's current navigation stack.
-            // Pages may have been pushed (e.g., from OnNavigatedTo) before this fragment
-            // was created by ViewPager2. Those NavigationRequested events were lost because
-            // nobody was subscribed yet. Reconcile by reading the section's actual stack.
-            var initialStack = new List<IView> { _rootPage };
-            if (shellSection is not null && shellSection.CurrentItem == _shellContent)
-            {
-                var sectionStack = shellSection.Stack;
-                for (int i = 1; i < sectionStack.Count; i++)
-                {
-                    if (sectionStack[i] is not null)
-                    {
-                        initialStack.Add(sectionStack[i]);
-                    }
-                }
-            }
-
-            _stackNavigationManager.RequestNavigation(new NavigationRequest(initialStack, false));
-
-            // Intentionally clearing (not replaying) pending navigation requests.
-            //
-            // When HasNavHost is false, navigation requests are queued into _pendingNavigationRequests.
-            // However, ConnectAndInitialize() builds the initial navigation stack directly from
-            // shellSection.Stack (see above), which already includes any pages pushed while queued.
-            //
-            // Replaying the queue would cause duplicate page pushes because those pages are already
-            // in the stack. Additionally, queued requests have e.Task == null (the TCS is only created
-            // after the HasNavHost check at line ~1043), so there is no hanging Task to worry about —
-            // the caller (ShellSection.OnPushAsync) safely handles null Task.
-            _pendingNavigationRequests.Clear();
-        }
-
-        void OnNavigationRequested(object? sender, NavigationRequestedEventArgs e)
-        {
-            // Only handle navigation for THIS ShellContent
-            // Note: Don't check IsVisible - ViewPager2 may report pages as not visible
-            // even when they are the current item
-            if (_stackNavigationManager is null)
-            {
-                return;
-            }
-
-            var shellSection = _shellContent?.Parent as ShellSection;
-            if (shellSection is null || shellSection.CurrentItem != _shellContent)
-            {
-                return;
-            }
-
-            // Check if StackNavigationManager is ready (has NavHost)
-            if (!_stackNavigationManager.HasNavHost)
-            {
-                // Queue the navigation request to be processed after initialization
-                _pendingNavigationRequests.Enqueue(e);
-                return;
-            }
-
-            // Create a TaskCompletionSource to serialize navigation requests.
-            // ShellSection.OnPushAsync awaits e.Task — without this, multiple pushes
-            // fire in rapid succession and BuildNavigationStack reads stale NavigationStack,
-            // causing intermediate pages to be lost from the stack.
-            //
-            // Skip TCS when StackNavigationManager is already navigating (e.g., initial setup
-            // from ConnectAndInitialize). In that case, the push will be queued internally by
-            // StackNavigationManager and processed after the current navigation completes via
-            // ProcessNavigationQueue. Creating a TCS here would block Shell.GoToAsync because
-            // the initial navigation's NavigationFinished would complete the wrong TCS.
-            if (!_stackNavigationManager.IsNavigating)
-            {
-                var tcs = new System.Threading.Tasks.TaskCompletionSource<bool>();
-                _navigationTaskCompletionSource = tcs;
-                e.Task = tcs.Task;
-            }
-
-            var requestedStack = BuildNavigationStack(e);
-            if (requestedStack is not null && requestedStack.Count > 0)
-            {
-                _stackNavigationManager.RequestNavigation(new NavigationRequest(requestedStack, e.Animated));
-            }
-            else
-            {
-                // Nothing to navigate — complete immediately
-                var pendingTcs = _navigationTaskCompletionSource;
-                _navigationTaskCompletionSource = null;
-                pendingTcs?.TrySetResult(true);
-            }
-        }
-
-        System.Threading.Tasks.TaskCompletionSource<bool>? _navigationTaskCompletionSource;
-        readonly Queue<NavigationRequestedEventArgs> _pendingNavigationRequests = new Queue<NavigationRequestedEventArgs>();
-
-        List<IView> BuildNavigationStack(NavigationRequestedEventArgs e)
-        {
-            var currentStack = _stackNavigationManager?.NavigationStack ?? new List<IView>();
-
-            switch (e.RequestType)
-            {
-                case NavigationRequestType.Push:
-                    var pushStack = new List<IView>(currentStack);
-                    if (e.Page is not null)
-                    {
-                        pushStack.Add(e.Page);
-                    }
-                    return pushStack;
-
-                case NavigationRequestType.Pop:
-                    if (currentStack.Count > 1)
-                    {
-                        var popStack = new List<IView>(currentStack);
-                        popStack.RemoveAt(popStack.Count - 1);
-                        return popStack;
-                    }
-                    return currentStack.ToList();
-
-                case NavigationRequestType.PopToRoot:
-                    if (currentStack.Count > 0)
-                    {
-                        return new List<IView> { currentStack[0] };
-                    }
-                    break;
-
-                case NavigationRequestType.Insert:
-                case NavigationRequestType.Remove:
-                    var section = _shellContent?.Parent as ShellSection;
-                    if (section is not null)
-                    {
-                        var resultStack = new List<IView>();
-                        foreach (var page in section.Stack)
-                        {
-                            resultStack.Add(page ?? _rootPage!);
-                        }
-                        return resultStack;
-                    }
-                    break;
-            }
-
-            return currentStack.ToList();
-        }
-
-        void IStackNavigation.RequestNavigation(NavigationRequest request)
-        {
-            _stackNavigationManager?.RequestNavigation(request);
-        }
-
-        // IStackNavigation.NavigationFinished - delegates to the internal method
-        void IStackNavigation.NavigationFinished(IReadOnlyList<IView> newStack)
-        {
-            OnNavigationFinished(newStack);
-        }
-
-        /// <summary>
-        /// Called by ShellContentStackNavigationView when navigation completes.
-        /// Updates toolbar and tab visibility based on the new navigation stack.
-        /// </summary>
-        internal void OnNavigationFinished(IReadOnlyList<IView> newStack)
-        {
-            // Complete the pending navigation task so ShellSection.OnPushAsync can proceed
-            var tcs = _navigationTaskCompletionSource;
-            _navigationTaskCompletionSource = null;
-            tcs?.TrySetResult(true);
-
-            if (!IsVisible || _handler is null)
-            {
-                return;
-            }
-
-            // Check if this content is currently active
-            var shellSection = _shellContent?.Parent as ShellSection;
-            var shellItem = shellSection?.Parent as ShellItem;
-            var shell = shellItem?.Parent as Shell;
-
-            if (shellSection is null || shellItem is null || shell is null)
-            {
-                return;
-            }
-
-            if (shell.CurrentItem != shellItem ||
-                shellItem.CurrentItem != shellSection ||
-                shellSection.CurrentItem != _shellContent)
-            {
-                return;
-            }
-
-            // Update toolbar
-            if (newStack.Count > 0 && newStack[newStack.Count - 1] is Page currentPage)
-            {
-                var toolbarTracker = _handler.ToolbarTracker;
-                var shellToolbar = _handler.ShellToolbar;
-                if (toolbarTracker is not null && shellToolbar is not null)
-                {
-                    toolbarTracker.CanNavigateBack = newStack.Count > 1;
-
-                    if (toolbarTracker.Page != currentPage)
-                    {
-                        toolbarTracker.Page = currentPage;
-                    }
-
-                    shellToolbar.Handler?.UpdateValue(nameof(IToolbar.BackButtonVisible));
-
-                    if (shell.Toolbar is ShellToolbar st)
-                    {
-                        st.ApplyChanges();
-                    }
-
-                    ((IShellController)shell).AppearanceChanged(currentPage, false);
-                }
-            }
-
-            // Hide/show content tabs based on navigation depth.
-            // Push beyond root → hide top tabs; pop to root → show top tabs.
-            // Uses PlaceTopTabs/RemoveTopTabs (fragment add/remove in NRM slot).
-            var shouldShowTabs = newStack.Count == 1 && ((IShellSectionController)shellSection).GetItems().Count > 1;
-
-            if (shouldShowTabs)
-            {
-                _handler.PlaceTopTabs();
-            }
-            else
-            {
-                _handler.RemoveTopTabs();
-            }
-        }
-
-        internal static void CloseActiveActionModes(AView? view)
-        {
-            if (view is global::Android.Widget.ListView listView)
-            {
-                // Unwrap HeaderViewListAdapter — ListViewRenderer adds header/footer views
-                var adapter = listView.Adapter;
-                if (adapter is HeaderViewListAdapter headerAdapter)
-                {
-                    adapter = headerAdapter.WrappedAdapter;
-                }
-
-                if (adapter is Compatibility.CellAdapter cellAdapter)
-                {
-                    cellAdapter.CloseContextActions();
-                }
-                return;
-            }
-
-            if (view is ViewGroup viewGroup)
-            {
-                for (int i = 0; i < viewGroup.ChildCount; i++)
-                {
-                    CloseActiveActionModes(viewGroup.GetChildAt(i));
-                }
-            }
-        }
-
-        public override void OnDestroyView()
-        {
-            // Disconnect StackNavigationManager IMMEDIATELY when the fragment view is destroyed.
-            // This removes the ViewAttachedToWindow listener before ViewPager2 can recycle/re-attach
-            // the view, preventing IllegalStateException from accessing Fragment on a destroyed view.
-            _navigationContainer?.ViewAttachedToWindow -= OnNavigationContainerAttached;
-
-            _stackNavigationManager?.Disconnect();
-
-            if (_subscribedToNavigationRequested)
-            {
-                var shellSection = _shellContent?.Parent as ShellSection;
-                if (shellSection is not null)
-                {
-                    ((IShellSectionController)shellSection).NavigationRequested -= OnNavigationRequested;
-                }
-                _subscribedToNavigationRequested = false;
-            }
-
-            base.OnDestroyView();
-        }
-
-        protected override void Dispose(bool disposing)
-        {
-            if (disposing)
-            {
-                _navigationContainer?.ViewAttachedToWindow -= OnNavigationContainerAttached;
-
-                if (_subscribedToNavigationRequested)
-                {
-                    var shellSection = _shellContent?.Parent as ShellSection;
-                    if (shellSection is not null)
-                    {
-                        ((IShellSectionController)shellSection).NavigationRequested -= OnNavigationRequested;
-                    }
-                    _subscribedToNavigationRequested = false;
-                }
-
-                _pendingNavigationRequests.Clear();
-                _navigationViewAdapter = null;
-                _stackNavigationManager?.Disconnect();
-                _stackNavigationManager = null;
-                _navigationContainer = null;
-                _rootPage = null;
-            }
-            base.Dispose(disposing);
-        }
-    }
-
-    /// <summary>
-    /// Adapter that wraps a Page (which is an IView) and adds IStackNavigation behavior
-    /// by delegating NavigationFinished to the owning fragment.
-    /// This is a cleaner approach than having the Fragment implement IStackNavigationView
-    /// with 40+ stub properties.
-    /// </summary>
-    internal class ShellContentStackNavigationView : IStackNavigationView
-    {
-        readonly Page _page;
-        readonly ShellContentNavigationFragment _fragment;
-
-        public ShellContentStackNavigationView(Page page, ShellContentNavigationFragment fragment)
-        {
-            _page = page ?? throw new ArgumentNullException(nameof(page));
-            _fragment = fragment ?? throw new ArgumentNullException(nameof(fragment));
-        }
-
-        // IStackNavigation - delegate to fragment
-        void IStackNavigation.RequestNavigation(NavigationRequest request)
-        {
-            // Navigation requests go through the fragment
-            ((IStackNavigation)_fragment).RequestNavigation(request);
-        }
-
-        // IStackNavigationView - delegate to fragment for the callback
-        public void NavigationFinished(IReadOnlyList<IView> newStack)
-        {
-            _fragment.OnNavigationFinished(newStack);
-        }
-
-        // IView - delegate all properties to the underlying Page
-        public Size Arrange(Rect bounds) => ((IView)_page).Arrange(bounds);
-        public Size Measure(double widthConstraint, double heightConstraint) => ((IView)_page).Measure(widthConstraint, heightConstraint);
-        public void InvalidateMeasure() => ((IView)_page).InvalidateMeasure();
-        public void InvalidateArrange() => ((IView)_page).InvalidateArrange();
-        public bool Focus() => ((IView)_page).Focus();
-        public void Unfocus() => ((IView)_page).Unfocus();
-
-        public string AutomationId => ((IView)_page).AutomationId;
-        public FlowDirection FlowDirection => ((IView)_page).FlowDirection;
-        public Primitives.LayoutAlignment HorizontalLayoutAlignment => ((IView)_page).HorizontalLayoutAlignment;
-        public Primitives.LayoutAlignment VerticalLayoutAlignment => ((IView)_page).VerticalLayoutAlignment;
-        public Semantics? Semantics => ((IView)_page).Semantics;
-        public IShape? Clip => ((IView)_page).Clip;
-        public IShadow? Shadow => ((IView)_page).Shadow;
-        public bool IsEnabled => ((IView)_page).IsEnabled;
-        public bool IsFocused { get => ((IView)_page).IsFocused; set => ((IView)_page).IsFocused = value; }
-        public Visibility Visibility => ((IView)_page).Visibility;
-        public double Opacity => ((IView)_page).Opacity;
-        public Paint? Background => ((IView)_page).Background;
-        public Rect Frame { get => ((IView)_page).Frame; set => ((IView)_page).Frame = value; }
-        public double Width => ((IView)_page).Width;
-        public double MinimumWidth => ((IView)_page).MinimumWidth;
-        public double MaximumWidth => ((IView)_page).MaximumWidth;
-        public double Height => ((IView)_page).Height;
-        public double MinimumHeight => ((IView)_page).MinimumHeight;
-        public double MaximumHeight => ((IView)_page).MaximumHeight;
-        public Thickness Margin => ((IView)_page).Margin;
-        public Size DesiredSize => ((IView)_page).DesiredSize;
-        public int ZIndex => ((IView)_page).ZIndex;
-        public IViewHandler? Handler { get => _page.Handler; set => _page.Handler = value; }
-        public bool InputTransparent => ((IView)_page).InputTransparent;
-        IElementHandler? IElement.Handler { get => _page.Handler; set => _page.Handler = value as IViewHandler; }
-        public IElement? Parent => ((IElement)_page).Parent;
-        public double TranslationX => ((IView)_page).TranslationX;
-        public double TranslationY => ((IView)_page).TranslationY;
-        public double Scale => ((IView)_page).Scale;
-        public double ScaleX => ((IView)_page).ScaleX;
-        public double ScaleY => ((IView)_page).ScaleY;
-        public double Rotation => ((IView)_page).Rotation;
-        public double RotationX => ((IView)_page).RotationX;
-        public double RotationY => ((IView)_page).RotationY;
-        public double AnchorX => ((IView)_page).AnchorX;
-        public double AnchorY => ((IView)_page).AnchorY;
-    }
-
-    /// <summary>
     /// TabLayout configuration strategy for ShellContent tabs
     /// </summary>
     internal class ShellTabConfigurationStrategy : Java.Lang.Object, TabLayoutMediator.ITabConfigurationStrategy
@@ -1465,108 +949,27 @@ namespace Microsoft.Maui.Controls.Handlers
             _wrapperFragment?.Dispose();
             _wrapperFragment = null;
         }
+    }
 
-        /// <summary>
-        /// Simple wrapper fragment - just returns the handler's unified view.
-        /// </summary>
-        class ShellSectionWrapperFragment : Fragment
+    /// <summary>
+    /// StackNavigationManager subclass that delegates animation creation to the ShellSectionHandler,
+    /// enabling custom page transition animations in Shell via protected virtual override.
+    /// </summary>
+    internal class ShellStackNavigationManager : StackNavigationManager
+    {
+        readonly ShellSectionHandler _handler;
+
+        public ShellStackNavigationManager(IMauiContext mauiContext, ShellSectionHandler handler)
+            : base(mauiContext)
         {
-            readonly ShellSectionHandler? _handler;
-            AView? _view;
+            _handler = handler;
+        }
 
-            // Default constructor required by Android's FragmentManager for fragment restoration
-            public ShellSectionWrapperFragment()
-            {
-                _handler = null;
-            }
-
-            public ShellSectionWrapperFragment(ShellSectionHandler handler)
-            {
-                _handler = handler;
-                _handler.SetParentFragment(this);
-            }
-
-            public override void OnCreate(Bundle? savedInstanceState)
-            {
-                // Always pass null to prevent restoring stale child fragment state.
-                // OffscreenPageLimit keeps fragments alive so restoration shouldn't occur,
-                // but this is defense-in-depth.
-                base.OnCreate(null);
-            }
-
-            public override AView OnCreateView(LayoutInflater inflater, ViewGroup? container, Bundle? savedInstanceState)
-            {
-                // If restored without proper handler reference, return empty view
-                if (_handler is null)
-                {
-                    return new FrameLayout(inflater.Context!);
-                }
-                if (_view is null)
-                {
-                    _view = _handler.PlatformView ?? _handler.ToPlatform();
-                }
-
-                // Remove from parent if it has one (fragment recreation scenario)
-                if (_view.Parent is ViewGroup parent)
-                {
-                    parent.RemoveView(_view);
-                }
-
-                return _view;
-            }
-
-            public override void OnViewCreated(AView view, Bundle? savedInstanceState)
-            {
-                base.OnViewCreated(view, savedInstanceState);
-
-                // Setup adapter now that fragment is attached
-                _handler?.SetupViewPagerAdapter();
-            }
-
-            public override void OnResume()
-            {
-                base.OnResume();
-
-                if (_handler is null || _handler.VirtualView is null)
-                {
-                    return;
-                }
-
-                if (!_handler.IsCurrentlyActiveSection())
-                {
-                    return;
-                }
-
-                var shell = _handler.VirtualView.FindParentOfType<Shell>();
-                var currentContent = _handler.VirtualView.CurrentItem;
-
-                if (shell is null || currentContent is null)
-                {
-                    return;
-                }
-
-                var page = ((IShellContentController)currentContent).GetOrCreateContent();
-
-                if (page is null)
-                {
-                    return;
-                }
-
-                // Update toolbar
-                var toolbarTracker = _handler.ToolbarTracker;
-                toolbarTracker?.Page = page;
-
-                ((IShellController)shell).AppearanceChanged(page, false);
-            }
-
-            protected override void Dispose(bool disposing)
-            {
-                if (disposing)
-                {
-                    _view = null;
-                }
-                base.Dispose(disposing);
-            }
+        public override AAnimation? OnCreateNavigationAnimation(
+            Context context, bool isPopping, bool enter)
+        {
+            return _handler.InvokeOnCreateNavigationAnimation(context, isPopping, enter)
+                ?? base.OnCreateNavigationAnimation(context, isPopping, enter);
         }
     }
 }
