+++ src/Core/src/Platform/Android/Navigation/NavigationViewFragment.cs
@@ -96,54 +96,17 @@ namespace Microsoft.Maui.Platform
 
 		public override Animation OnCreateAnimation(int transit, bool enter, int nextAnim)
 		{
-			int id = 0;
-
-			Animation? returnValue;
-
 			// This means the operation currently being processed shouldn't be animated
 			// This will happen if a user inserts or removes a root page
 			if (NavigationManager.IsPopping == null || !NavigationManager.IsAnimated)
 			{
-				returnValue = null;
-			}
-			else
-			{
-				// Once we have Function Mappers figured out all of this code can
-				// move to a function mapper as a way to customize animations from code
-				if (NavigationManager.IsPopping.Value)
-				{
-					if (!enter)
-					{
-						id = Resource.Animation.nav_default_pop_exit_anim;
-					}
-					else
-					{
-						id = Resource.Animation.nav_default_pop_enter_anim;
-					}
-				}
-				else
-				{
-					if (enter)
-					{
-						id = Resource.Animation.nav_default_enter_anim;
-					}
-					else
-					{
-						id = Resource.Animation.nav_default_exit_anim;
-					}
-				}
-
-				if (id > 0)
-				{
-					returnValue = AnimationUtils.LoadAnimation(Context, id);
-				}
-				else
-				{
-					returnValue = base.OnCreateAnimation(transit, enter, id);
-				}
+				return null!;
 			}
 
-			return returnValue!;
+			// Delegate to StackNavigationManager which can be overridden for custom animations
+			var animation = NavigationManager.OnCreateNavigationAnimation(Context!, NavigationManager.IsPopping.Value, enter);
+
+			return (animation ?? base.OnCreateAnimation(transit, enter, 0))!;
 		}
 	}
 }
